Create Icon subset for Beacon
