+++ src/components/Icon/icons.all.js
@@ -1,0 +1,408 @@
+import activity from '../../icons/activity.svg'
+import alert from '../../icons/alert.svg'
+import alignCenter from '../../icons/align-center.svg'
+import alignLeft from '../../icons/align-left.svg'
+import alignRight from '../../icons/align-right.svg'
+import animalsAndNature from '../../icons/animals-and-nature.svg'
+import arrowDownHuge from '../../icons/arrow-down-huge.svg'
+import arrowLeft from '../../icons/arrow-left.svg'
+import arrowLeftSmall from '../../icons/arrow-left-small.svg'
+import arrowMerge from '../../icons/arrow-merge.svg'
+import arrowRight from '../../icons/arrow-right.svg'
+import arrowRightSmall from '../../icons/arrow-right-small.svg'
+import arrowSwerve from '../../icons/arrow-swerve.svg'
+import arrowTwoWay from '../../icons/arrow-two-way.svg'
+import arrowUpHuge from '../../icons/arrow-up-huge.svg'
+import assign from '../../icons/assign.svg'
+import assigned from '../../icons/assigned.svg'
+import at from '../../icons/at.svg'
+import attach from '../../icons/attach.svg'
+import attachLarge from '../../icons/attach-large.svg'
+import attachment from '../../icons/attachment.svg'
+import attachSmall from '../../icons/attach-small.svg'
+import bcc from '../../icons/bcc.svg'
+import bell from '../../icons/bell.svg'
+import bellSmall from '../../icons/bell-small.svg'
+import bin from '../../icons/bin.svg'
+import bold from '../../icons/bold.svg'
+import bubbleDots from '../../icons/bubble-dots.svg'
+import bubbleDotsDuplicate from '../../icons/bubble-dots-duplicate.svg'
+import bubbleDotsPlus from '../../icons/bubble-dots-plus.svg'
+import bubbleTick from '../../icons/bubble-tick.svg'
+import buoy from '../../icons/buoy.svg'
+import calendar from '../../icons/calendar.svg'
+import call from '../../icons/call.svg'
+import callLarge from '../../icons/call-large.svg'
+import camera from '../../icons/camera.svg'
+import caretDown from '../../icons/caret-down.svg'
+import caretLeft from '../../icons/caret-left.svg'
+import caretRight from '../../icons/caret-right.svg'
+import caretUp from '../../icons/caret-up.svg'
+import chat from '../../icons/chat.svg'
+import chatActive from '../../icons/chat-active.svg'
+import chatSmall from '../../icons/chat-small.svg'
+import chatSquare from '../../icons/chat-square.svg'
+import check from '../../icons/check.svg'
+import checkboxLarge from '../../icons/checkbox-large.svg'
+import checkboxSmall from '../../icons/checkbox-small.svg'
+import circleArrowDown from '../../icons/circle-arrow-down.svg'
+import circleArrowLeft from '../../icons/circle-arrow-left.svg'
+import circleArrowRight from '../../icons/circle-arrow-right.svg'
+import circleArrowUp from '../../icons/circle-arrow-up.svg'
+import circleCross from '../../icons/circle-cross.svg'
+import circleMinus from '../../icons/circle-minus.svg'
+import circlePlus from '../../icons/circle-plus.svg'
+import circleTick from '../../icons/circle-tick.svg'
+import clock from '../../icons/clock.svg'
+import clockLarge from '../../icons/clock-large.svg'
+import clockSmall from '../../icons/clock-small.svg'
+import closed from '../../icons/closed.svg'
+import cloud from '../../icons/cloud.svg'
+import cog from '../../icons/cog.svg'
+import collapse from '../../icons/collapse.svg'
+import compose from '../../icons/compose.svg'
+import connection from '../../icons/connection.svg'
+import copy from '../../icons/copy.svg'
+import crossLarge from '../../icons/cross-large.svg'
+import crossMedium from '../../icons/cross-medium.svg'
+import crossMediumAlt from '../../icons/cross-medium-alt.svg'
+import crossSmall from '../../icons/cross-small.svg'
+import cursor from '../../icons/cursor.svg'
+import dashboard from '../../icons/dashboard.svg'
+import document from '../../icons/document.svg'
+import documentIn from '../../icons/document-in.svg'
+import documentOut from '../../icons/document-out.svg'
+import dollar from '../../icons/dollar.svg'
+import drafts from '../../icons/drafts.svg'
+import drag from '../../icons/drag.svg'
+import dragHandle from '../../icons/drag-handle.svg'
+import duplicate from '../../icons/duplicate.svg'
+import ellipsis from '../../icons/ellipsis.svg'
+import ellipsisSmall from '../../icons/ellipsis-small.svg'
+import emoji from '../../icons/emoji.svg'
+import envelope from '../../icons/envelope.svg'
+import envelopeLarge from '../../icons/envelope-large.svg'
+import envelopeOpen from '../../icons/envelope-open.svg'
+import eye from '../../icons/eye.svg'
+import eyeClosed from '../../icons/eye-closed.svg'
+import fabAntenna from '../../icons/fab-antenna.svg'
+import fabBuoy from '../../icons/fab-buoy.svg'
+import fabChat from '../../icons/fab-chat.svg'
+import fabQuestion from '../../icons/fab-question.svg'
+import fabSearch from '../../icons/fab-search.svg'
+import fields from '../../icons/fields.svg'
+import flags from '../../icons/flags.svg'
+import folder from '../../icons/folder.svg'
+import folderCheck from '../../icons/folder-check.svg'
+import folderOpen from '../../icons/folder-open.svg'
+import folderOpenSmall from '../../icons/folder-open-small.svg'
+import foodAndDrink from '../../icons/food-and-drink.svg'
+import forward from '../../icons/forward.svg'
+import fullscreen from '../../icons/fullscreen.svg'
+import globe from '../../icons/globe.svg'
+import graph from '../../icons/graph.svg'
+import heart from '../../icons/heart.svg'
+import helpscoutLogo from '../../icons/helpscout-logo.svg'
+import house from '../../icons/house.svg'
+import hsLogo from '../../icons/hs-logo.svg'
+import hyphen from '../../icons/hyphen.svg'
+import image from '../../icons/image.svg'
+import imageAdd from '../../icons/image-add.svg'
+import imageLarge from '../../icons/image-large.svg'
+import imagePlus from '../../icons/image-plus.svg'
+import inbox from '../../icons/inbox.svg'
+import info from '../../icons/info.svg'
+import information from '../../icons/information.svg'
+import italic from '../../icons/italic.svg'
+import kebab from '../../icons/kebab.svg'
+import key from '../../icons/key.svg'
+import keyboard from '../../icons/keyboard.svg'
+import largeDoubleArrowLeft from '../../icons/large-double-arrow-left.svg'
+import largeDoubleArrowRight from '../../icons/large-double-arrow-right.svg'
+import largeSingleArrowLeft from '../../icons/large-single-arrow-left.svg'
+import largeSingleArrowRight from '../../icons/large-single-arrow-right.svg'
+import link from '../../icons/link.svg'
+import list from '../../icons/list.svg'
+import lockClosed from '../../icons/lock-closed.svg'
+import lockOpen from '../../icons/lock-open.svg'
+import magazine from '../../icons/magazine.svg'
+import markup from '../../icons/markup.svg'
+import meatball from '../../icons/meatball.svg'
+import menuSmall from '../../icons/menu-small.svg'
+import mine from '../../icons/mine.svg'
+import minimize from '../../icons/minimize.svg'
+import minus from '../../icons/minus.svg'
+import moon from '../../icons/moon.svg'
+import newConvo from '../../icons/new-convo.svg'
+import newWindow from '../../icons/new-window.svg'
+import newWindowLarge from '../../icons/new-window-large.svg'
+import note from '../../icons/note.svg'
+import objects from '../../icons/objects.svg'
+import offline from '../../icons/offline.svg'
+import paperPlane from '../../icons/paper-plane.svg'
+import pause from '../../icons/pause.svg'
+import pencilSmall from '../../icons/pencil-small.svg'
+import period from '../../icons/period.svg'
+import personMultiple from '../../icons/person-multiple.svg'
+import personPlus from '../../icons/person-plus.svg'
+import personTick from '../../icons/person-tick.svg'
+import personX from '../../icons/person-x.svg'
+import playVideo from '../../icons/play-video.svg'
+import plusLarge from '../../icons/plus-large.svg'
+import plusMedium from '../../icons/plus-medium.svg'
+import plusSmall from '../../icons/plus-small.svg'
+import power from '../../icons/power.svg'
+import printer from '../../icons/printer.svg'
+import quote from '../../icons/quote.svg'
+import recordVideo from '../../icons/record-video.svg'
+import refresh from '../../icons/refresh.svg'
+import reply from '../../icons/reply.svg'
+import rosette from '../../icons/rosette.svg'
+import screen from '../../icons/screen.svg'
+import search from '../../icons/search.svg'
+import searchMedium from '../../icons/search-medium.svg'
+import searchSmall from '../../icons/search-small.svg'
+import shoppingCart from '../../icons/shopping-cart.svg'
+import sign from '../../icons/sign.svg'
+import smallPaperclip from '../../icons/small-paperclip.svg'
+import socialFacebook from '../../icons/social-facebook.svg'
+import socialTwitter from '../../icons/social-twitter.svg'
+import spam from '../../icons/spam.svg'
+import spinner from '../../icons/spinner.svg'
+import star from '../../icons/star.svg'
+import starSmall from '../../icons/star-small.svg'
+import status from '../../icons/status.svg'
+import stop from '../../icons/stop.svg'
+import suitcase from '../../icons/suitcase.svg'
+import symbols from '../../icons/symbols.svg'
+import table from '../../icons/table.svg'
+import tag from '../../icons/tag.svg'
+import text from '../../icons/text.svg'
+import tickLarge from '../../icons/tick-large.svg'
+import tickSmall from '../../icons/tick-small.svg'
+import toggle from '../../icons/toggle.svg'
+import trash from '../../icons/trash.svg'
+import travelAndPlaces from '../../icons/travel-and-places.svg'
+import trayLarge from '../../icons/tray-large.svg'
+import traySmall from '../../icons/tray-small.svg'
+import traySuccess from '../../icons/tray-success.svg'
+import trophy from '../../icons/trophy.svg'
+import unassigned from '../../icons/unassigned.svg'
+import underline from '../../icons/underline.svg'
+import user from '../../icons/user.svg'
+import userX from '../../icons/user-x.svg'
+import workflow from '../../icons/workflow.svg'
+import workflowSmall from '../../icons/workflow-small.svg'
+import xLargeSingleArrowLeft from '../../icons/x-large-single-arrow-left.svg'
+import xLargeSingleArrowRight from '../../icons/x-large-single-arrow-right.svg'
+
+const ICONS = {
+  activity,
+  alert,
+  'align-center': alignCenter,
+  'align-left': alignLeft,
+  'align-right': alignRight,
+  'animals-and-nature': animalsAndNature,
+  'arrow-down-huge': arrowDownHuge,
+  'arrow-left-double-large': largeDoubleArrowLeft,
+  'arrow-left-single-large': largeSingleArrowLeft,
+  'arrow-left-small': arrowLeftSmall,
+  'arrow-left': arrowLeft,
+  'arrow-merge': arrowMerge,
+  'arrow-right-double-large': largeDoubleArrowRight,
+  'arrow-right-single-large': largeSingleArrowRight,
+  'arrow-right-small': arrowRightSmall,
+  'arrow-right': arrowRight,
+  'arrow-swerve': arrowSwerve,
+  'arrow-two-way': arrowTwoWay,
+  'arrow-up-huge': arrowUpHuge,
+  assign,
+  assigned,
+  at,
+  'attach-large': attachLarge,
+  'attach-small': attachSmall,
+  attach,
+  attachment,
+  bcc,
+  'bell-small': bellSmall,
+  bell,
+  bin: trash,
+  bold,
+  'bubble-dots-duplicate': bubbleDotsDuplicate,
+  'bubble-dots-plus': bubbleDotsPlus,
+  'bubble-dots': bubbleDots,
+  'bubble-tick': bubbleTick,
+  buoy,
+  calendar,
+  'call-large': callLarge,
+  call,
+  camera,
+  'caret-down': caretDown,
+  'caret-left': caretLeft,
+  'caret-right': caretRight,
+  'caret-up': caretUp,
+  'chat-active': chatActive,
+  'chat-small': chatSmall,
+  'chat-square': chatSquare,
+  chat,
+  check,
+  'checkbox-large': checkboxLarge,
+  'checkbox-small': checkboxSmall,
+  'circle-arrow-down': circleArrowDown,
+  'circle-arrow-left': circleArrowLeft,
+  'circle-arrow-right': circleArrowRight,
+  'circle-arrow-up': circleArrowUp,
+  'circle-cross': circleCross,
+  'circle-minus': circleMinus,
+  'circle-plus': circlePlus,
+  'circle-tick': circleTick,
+  'clock-large': clockLarge,
+  'clock-small': clockSmall,
+  clock,
+  clockLarge,
+  closed,
+  cloud,
+  cog,
+  collapse,
+  compose,
+  connection,
+  copy,
+  'cross-large': crossLarge,
+  'cross-medium-alt': crossMediumAlt,
+  'cross-medium': crossMedium,
+  'cross-small': crossSmall,
+  crossLarge,
+  cursor,
+  dashboard,
+  'document-in': documentIn,
+  'document-out': documentOut,
+  document,
+  dollar,
+  drafts,
+  'drag-handle': dragHandle,
+  drag,
+  duplicate,
+  'ellipsis-small': ellipsisSmall,
+  ellipsis,
+  emoji,
+  'envelope-large': envelopeLarge,
+  'envelope-open': envelopeOpen,
+  envelope,
+  'eye-closed': eyeClosed,
+  eye,
+  'fab-antenna': fabAntenna,
+  'fab-buoy': fabBuoy,
+  'fab-chat': fabChat,
+  'fab-question': fabQuestion,
+  'fab-search': fabSearch,
+  fields,
+  flags,
+  'folder-check': folderCheck,
+  'folder-open-small': folderOpenSmall,
+  'folder-open': folderOpen,
+  folder,
+  'food-and-drink': foodAndDrink,
+  forward,
+  fullscreen,
+  globe,
+  graph,
+  heart,
+  'helpscout-logo': helpscoutLogo,
+  house,
+  'hs-logo': hsLogo,
+  hyphen,
+  'image-add': imageAdd,
+  'image-large': imageLarge,
+  'image-plus': imagePlus,
+  image,
+  inbox,
+  info,
+  information,
+  italic,
+  kebab,
+  key,
+  keyboard,
+  'large-double-arrow-left': largeDoubleArrowLeft,
+  'large-double-arrow-right': largeDoubleArrowRight,
+  'large-single-arrow-left': largeSingleArrowLeft,
+  'large-single-arrow-right': largeSingleArrowRight,
+  link,
+  list,
+  'lock-closed': lockClosed,
+  'lock-open': lockOpen,
+  magazine,
+  markup,
+  meatball,
+  'menu-small': menuSmall,
+  mine,
+  minimize,
+  minus,
+  moon,
+  'new-convo': newConvo,
+  'new-window-large': newWindowLarge,
+  'new-window': newWindow,
+  note,
+  objects,
+  offline,
+  'option-dots': meatball,
+  'paper-plane': paperPlane,
+  pause,
+  'pencil-small': pencilSmall,
+  period,
+  'person-multiple': personMultiple,
+  'person-plus': personPlus,
+  'person-tick': personTick,
+  'person-x': personX,
+  'play-video': playVideo,
+  'plus-large': plusLarge,
+  'plus-medium': plusMedium,
+  'plus-small': plusSmall,
+  plus: plusLarge,
+  power,
+  printer,
+  quote,
+  'record-video': recordVideo,
+  refresh,
+  reply,
+  rosette,
+  screen,
+  'search-medium': searchMedium,
+  'search-small': searchSmall,
+  search,
+  'shopping-cart': shoppingCart,
+  sign,
+  'small-paperclip': smallPaperclip,
+  'social-facebook': socialFacebook,
+  'social-twitter': socialTwitter,
+  spam,
+  spinner,
+  'star-small': starSmall,
+  star,
+  status,
+  stop,
+  suitcase,
+  symbols,
+  table,
+  tag,
+  text,
+  'tick-large': tickLarge,
+  'tick-small': tickSmall,
+  tick: tickLarge,
+  toggle,
+  trash,
+  'travel-and-places': travelAndPlaces,
+  'tray-large': trayLarge,
+  'tray-small': traySmall,
+  'tray-success': traySuccess,
+  trophy,
+  unassigned,
+  underline,
+  'user-x': userX,
+  user,
+  'workflow-small': workflowSmall,
+  workflow,
+  'x-large-single-arrow-left': xLargeSingleArrowLeft,
+  'x-large-single-arrow-right': xLargeSingleArrowRight,
+}
+
+export default ICONS
